refactor(settings): split widget primitives into settings_widgets.py
Move the theme constants, the NSColor/NSFont helpers, the layer-drawn
NSView/NSButton/NSTextFieldCell subclasses (_PinkBackgroundView,
_SettingsNSWindow, _PaddedTextFieldCell, _PillButton, _GhostPopUp,
_SegButton, _GhostSegment, _CardView, _DividerView), and the factory
functions (_label, _button, _text_field, _checkbox, _popup, _segmented,
_setting_row, _field_row, _card, etc.) into a new module so the
controllers stop sharing a 3464-line file with their building blocks.

settings_window.py drops from 3464 to 2530 lines; settings_widgets.py
is 1013 lines of pure UI primitives with no daemon, config, or persona
dependencies. Behaviour is unchanged — both SettingsController and
_OnboardingController stay in settings_window.py, the public entry
points (show, show_onboarding) still resolve, url_scheme.py's private
imports (_OnboardingController, _self_test_managed_async) still exist
at their original paths, and all 352 tests + lint pass.
